requirements.txt
Django==2.1.1
PyMySQL==0.9.2
django-bootstrap4
django-multiselectfield
